Rename `storage::batch` to `storage::arrow`
