+++ packages/engine/lib/storage/src/arrow/mod.rs
@@ -1,4 +1,4 @@
-pub mod change;
+mod change;
 pub mod flush;
 pub mod ipc;
 pub mod load;
@@ -8,6 +8,7 @@ mod conversion;
 
 pub use self::{
     arrow_batch::ArrowBatch,
+    change::ColumnChange,
     conversion::{
         col_to_json_vals, json_utf8_json_vals, json_vals_to_any_type_col, json_vals_to_bool,
         json_vals_to_col, json_vals_to_primitive, json_vals_to_utf8, new_buffer,
