+++ packages/engine/src/datastore/batch/agent.rs
@@ -16,7 +16,7 @@ use arrow::{
     record_batch::RecordBatch,
 };
 use storage::{
-    batch::{
+    arrow::{
         flush::GrowableBatch,
         ipc::{record_batch_data_to_bytes_owned_unchecked, simulate_record_batch_to_bytes},
         ArrowBatch,
