+++ packages/engine/src/datastore/batch/message.rs
@@ -12,7 +12,7 @@ use arrow::{
     record_batch::RecordBatch,
 };
 use storage::{
-    batch::{
+    arrow::{
         flush::GrowableBatch,
         ipc::{record_batch_data_to_bytes_owned_unchecked, simulate_record_batch_to_bytes},
         ArrowBatch,
