+++ packages/engine/src/simulation/package/state/packages/behavior_execution/mod.rs
@@ -23,7 +23,7 @@ use crate::{
                     fields::{BEHAVIOR_IDS_FIELD_NAME, BEHAVIOR_INDEX_FIELD_NAME},
                     tasks::ExecuteBehaviorsTask,
                 },
-                Arc, ColumnChange, DatastoreResult, Error, ExperimentConfig, FieldSpecMapAccessor,
+                Arc, DatastoreResult, Error, ExperimentConfig, FieldSpecMapAccessor,
                 GetWorkerExpStartMsg, GetWorkerSimStartMsg, Globals, IntoArrowChange, Name,
                 Package, PackageComms, PackageCreator, Result, RootFieldSpec, RootFieldSpecCreator,
                 SimRunConfig, Span, State, StateColumn, StateTask,
