+++ packages/engine/src/simulation/package/state/packages/behavior_execution/reset_index_col.rs
@@ -1,8 +1,8 @@
 use arrow::{array::ArrayData, datatypes::DataType};
-use storage::batch::new_buffer;
+use storage::arrow::{new_buffer, ColumnChange};
 
 use crate::simulation::package::state::packages::behavior_execution::{
-    BehaviorIndexInnerDataType, ColumnChange, DatastoreResult, IntoArrowChange, Result, StateColumn,
+    BehaviorIndexInnerDataType, DatastoreResult, IntoArrowChange, Result, StateColumn,
 };
 
 pub fn reset_index_col(behavior_index_col_index: usize) -> Result<StateColumn> {
